// Native/form/forms/types/types.ts
export interface FormData {
    productName: string;
    description: string;
    price: string;
    quantity: string;
    date: string;
    category: string;
    condition: string;
    email: string;
    pincode: string;
    gstNumber: string;
    country: string;
    password: string;
    confirmPassword: string;
    phoneNumber: string;
    website: string;
}